requirements.txt
alembic==1.13.1
annotated-types
anyio
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.3
cffi==1.16.0
click
cryptography==42.0.8
fastapi
greenlet==3.0.3
h11
idna
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pycparser==2.22
pydantic
pydantic_core
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio
SQLAlchemy==2.0.31
starlette
typing_extensions
uvicorn
requests
fastapi-pagination==0.12.25